+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/extensions/OptionalExtensions.kt
@@ -9,5 +9,5 @@ import java.util.*
 inline fun <T> Optional<T>.orNull(): T? = this.orElse(null)
 
 /**从一个可空对象创建Optional对象。*/
-@Deprecated("It is not recommended to use Optional objects in Kotlin directly.")
-inline fun <T> T?.toOptional(): Optional<T> = Optional.ofNullable(this)
+@Deprecated("It is not recommended to use Optional objects in Kotlin directly.", level = DeprecationLevel.HIDDEN)
+inline fun <T> T.toOptional(): Optional<T> = Optional.ofNullable(this)